requirements.txt
python-telegram-bot==13.13
huggingface-hub==0.9.1
